Renamed PatcherUtils to PatcherUtil. Trying to make material settings render pipeline agnostic. Added some more patch routines
