--- VisualPinball.Unity/VisualPinball.Unity.Patcher/Patcher/Common/PatcherUtils.cs
@@ -1,89 +1,0 @@
-﻿using System;
-using UnityEngine;
-using VisualPinball.Unity.Patcher.Matcher;
-using RenderPipeline = VisualPinball.Unity.Patcher.Matcher.RenderPipeline;
-
-namespace VisualPinball.Unity.Patcher
-{
-	static class PatcherUtils
-	{
-		public static void Hide(GameObject gameObject)
-		{
-			gameObject.SetActive(false);
-		}
-
-		public static void Reparent(GameObject child, GameObject parent)
-		{
-			var rot = child.transform.rotation;
-			var pos = child.transform.position;
-
-			// re-parent the child
-			child.transform.SetParent(parent.transform, false);
-
-			child.transform.rotation = rot;
-			child.transform.position = pos;
-		}
-
-		public static void SetOpaque(GameObject gameObject)
-		{
-			var material = gameObject.GetComponent<Renderer>().sharedMaterial;
-
-			switch (RenderPipeline.Current)
-			{
-				case RenderPipelineType.BuiltIn:
-					// TODO
-					break;
-
-				case RenderPipelineType.Hdrp:
-					material.SetFloat("_SurfaceType", 0);
-
-					material.SetFloat("_DstBlend", 0);
-					material.SetFloat("_ZWrite", 1);
-
-					material.DisableKeyword("_ALPHATEST_ON");
-					material.DisableKeyword("_SURFACE_TYPE_TRANSPARENT");
-					material.DisableKeyword("_BLENDMODE_PRE_MULTIPLY");
-					material.DisableKeyword("_BLENDMODE_PRESERVE_SPECULAR_LIGHTING");
-					break;
-
-				case RenderPipelineType.Urp:
-					// TODO
-					break;
-
-				default:
-					throw new ArgumentOutOfRangeException();
-			}
-		}
-
-		public static void SetDoubleSided(GameObject gameObject)
-		{
-
-			var material = gameObject.GetComponent<Renderer>().sharedMaterial;
-
-			switch (RenderPipeline.Current)
-			{
-				case RenderPipelineType.BuiltIn:
-					// TODO
-					break;
-
-				case RenderPipelineType.Hdrp:
-					material.EnableKeyword("_DOUBLESIDED_ON");
-					material.EnableKeyword("_NORMALMAP_TANGENT_SPACE");
-
-					material.SetInt("_DoubleSidedEnable", 1);
-					material.SetInt("_DoubleSidedNormalMode", 1);
-
-					material.SetInt("_CullMode", 0);
-					material.SetInt("_CullModeForward", 0);
-					break;
-
-				case RenderPipelineType.Urp:
-					// TODO
-					break;
-
-				default:
-					throw new ArgumentOutOfRangeException();
-			}
-		}
-	}
-}
